Sort the hashed email addresses before writing to output file
We need to sort the hashed email addresses instead of directly writing
them to the file. This helps us hide when a member was also added and
also makes lists intersections easier.

Note that this does result in an overhead since now we are dealing with
a list instead of just writing to the file.